Ridiculously large timeout (#588)
While investigating https://github.com/cirruslabs/cirrus-ci-docs/issues/1120 it appeared that occasional slowdowns happen for different reasons: network, load balancer, microservice scaling up and cold startup time, etc. Unfortunately, for some reason the call cancelation is not propogated to the backend properly, so it finishes assigning a task to a worker but worker never gets the assigment.

This timeout is our safeguard in case one of the upstream services like Cirrus CI and Cirrus Runners hangs indefinitely. But we've never seen such behaviour, so it make sense to increase this timeout to some very large value inversely proportional to a very low probability of an upstream hang.