Fixed missing dependencies in a complex `depends_on`/`matrix` case (#800)
* Fixed missing dependencies in a complex `depends_on`/`matrix` case

We did validation of unbalanced tasks first thing in dependency validation. But with a matrix is possible to get several tasks with the same name where some will be balanced and some not. We do matching by name though.

This change move the check after we set as many dependencies as we can. See the test case for more details.

* buf generate

* go generate